feat: 권한 관리

+++ src/main/java/com/changbi/tradeunion/boardforworkers/board/domain/Board.java
@@ -37,7 +37,7 @@ public class Board {
     private Integer depth;
 
     @Column
-    private Integer upperBoardId;
+    private Long upperBoardId;
 
     @Column
     private String boardPath;
